+++ mindstack_app/modules/gamification/logics/__init__.py
@@ -1,0 +1,1 @@
+"""Gamification Logic - Pure functions."""